scheduler/requirements.txt
aiohttp==3.11.18
asyncpg==0.30.0
redis==6.2.0
python-dotenv==0.21.0
aiofiles==24.1.0
msgpack==1.1.1
aiolimiter==1.2.1
